# config/userSwagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Edify lives",
        "description": "project description",
        "version": "1.0"
    },
    "produces": ["application/json"],
    "basePath": "/",
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    },
    "paths": {
...
        "/users/login": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "users-login-post",
                "tags": ["Application User"],
                "description": "",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Enter the email",
                        "required": true,
                        "type": "string",
                        "value": "vaneet@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Enter password",
                        "required": true,
                        "type": "string",
                        "value": "123456"
                    },
                    {
                        "name": "deviceToken",
                        "in": "formData",
                        "description": "Enter deviceToken",
                        "required": false,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceType",
                        "in": "formData",
                        "description": "Enter deviceType",
                        "enum": [1,2],
                        "required": false,
                        "type": "number",
                        "value": 1
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
}